Fix Kots existing cluster snapshot onboarding flow for LVP plugin
